Replace mt_rand and mcrypt_create_iv with random_bytes
mt_rand "must not be used for cryptographic purposes" according
to the docs, whereas the newer random_bytes function calls the
appropriate platform-specific API for cryptographically secure
random numbers.

mcrypt_create_iv is deprecated and removed, with random_bytes
being the replacement suggested in the PHP docs.

Additionally, by delegating to a protected method, the choice of
random number generator can be overridden if needed.